gitfs, wiki: Add Page Rename feature.
This commit also introduces limited Rename Tracking.

Rename Tracking works only under the certain conditions; The content of
the file needs not to be modified at the commit the file is renamed. If
you always edit all pages with only this application, it is no problem
because if you rename and edit a page, the rename and editing will be
commited separately. But if you push commits here from a repository
which is not managed by this application, Rename Tracking may not work
very well as git.