gitfs: Fix the problem with gitfs.log ignores limit
The problem was: While gitfs.log collect interest commits logs in the
entire history, it decreases query.limit for every interet commit and
if the limit is zero then stops. But sometimes query.limit goes down to
negative value and then gitfs.log cannot sense it's the time to stop.